Raise the fast Wald-Bikker ray intersection into separate code, make TriangleBVH use it.
git-svn-id: svn+ssh://jhoberock-dt/home/jhoberock/svnbase/jared/src/gotham@1453 3bb10773-34f4-0310-b88c-ec9d511ba0d9

+++ surfaces/Mesh.cpp
@@ -299,7 +299,7 @@ void Mesh
   ::getWaldBikkerData(const Triangle &tri,
                       WaldBikkerData &data) const
 {
-  size_t k;
+  unsigned int k;
   buildWaldBikkerIntersectionData<gpcpu::float3,float>(mPoints[tri[0]],
                                                        mPoints[tri[1]],
                                                        mPoints[tri[2]],
